requirements/common.txt
djangorestframework-jwt==1.11.0
django-environ==0.4.5
django-cors-headers==2.4.0
djoser==1.2.2

pytest==3.7.1
pytest-cov==2.5.1
pytest-django==3.4.1
coverage==4.5.1
